// paimon-web-ui/src/api/models/job/types/record.ts
/* Licensed to the Apache Software Foundation (ASF) under one
or more contributor license agreements.  See the NOTICE file
distributed with this work for additional information
regarding copyright ownership.  The ASF licenses this file
to you under the Apache License, Version 2.0 (the
"License"); you may not use this file except in compliance
with the License.  You may obtain a copy of the License at

  http://www.apache.org/licenses/LICENSE-2.0

  Unless required by applicable law or agreed to in writing,
software distributed under the License is distributed on an
"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
KIND, either express or implied.  See the License for the
specific language governing permissions and limitations
under the License. */

export interface Record {
  id: number
  name: string
  taskType: string
  isStreaming: boolean
  uid: number
  clusterId: number
  statements: string
  statementName?: string
}

export interface RecordNameParams {
  statementName?: string
  pageNum: number
  pageSize: number
}

export interface RecordDTO {
  statementName: string
  taskType: string
  isStreaming: boolean
  uid?: number
  clusterId?: number
  statements: string
}